Make sure locals generated by a frozen def inherit the right module slots
Summary:
When you see:

```
def f():
   def g():
...
```

The `g` needs to get the same module slots as the `f`. For non-frozen modules, we use `None` as the module slots, and we always set the slots to None, so that worked fine. If the module has been frozen we fill in the slots, so when a frozen function `f` creates a new `g` it wasn't getting the right slots. Fix that, which requires passing the `FrozenValue` representing the slots around a little bit more.

Reviewed By: abrassel

Differential Revision: D30345657

fbshipit-source-id: 1ca4abc2f2de584a3ce5b5c1942274d99c6ebbd4