Require the default for field has the right type
Summary: Previously you could define a record, with a field, and give the field a default value that did not match the annotated type of the field. That feels quite wrong so do `field(str.type, 42)`, since the field isn't always a string. Now catch that and raise an error.

Differential Revision: D30367412

fbshipit-source-id: 68174047dc40309334cf5cfb3acdd1a75399b086