Minor refactoring of starlark-derive
Summary:
Minor code cleanup: convert tuple to a struct with named fields.

To be able to introduce additional function attributes like `pure` or `spec_safe`.

At this moment, my experiments with `spec_safe` did not produce significant speed up, but submitting a diff to continue experimenting and not spending time on rebasing.

Reviewed By: ndmitchell

Differential Revision: D30722044

fbshipit-source-id: f63ec3d3c6391b0c66c27cbc0845ff29981e2537