Allow qualified names in #[starlark_module]
Summary: For `#[starlark_module]` you must take an argument of type `GlobalsBuilder`, but we don't care too much, so `starlark::environment::GlobalsBuilder` is fine, as is `foo::GlobalsBuilder`. To do that check we looked at the _first_ component of the type, which is a pure bug - it should have been the last. This will apply to other things like Option, Value etc, as it was always expected to.

Reviewed By: stepancheg

Differential Revision: D30724660

fbshipit-source-id: 81839f7e6e35d03170e5b2c8b5d0e602daab2dad