ExprEvalException
Summary:
`EvalException` is either error or flag to return, break or continue.

Which is misleading type when evaluating expressions: expressions cannot return, break or continue.

`ExprEvalException` is a new type which is wrapper for `anyhow::Error` which is an error returned from expression evaluation.

Additionally, `ExprEvalException` seems to be more efficient than `EvalException`: `EvalException` is two words, while `ExprEvalException` is one word.

See this example in compiled explorer: [link](https://rust.godbolt.org/z/3ds6W6dhE): `expr2` and `stmt2` functions are supposed to do the same, but `expr2` machine code is more efficient.

(`expr2` still does unnecessary work with `test` and `setne`, but at least it does not do branches and twice as short.)

Reviewed By: ndmitchell

Differential Revision: D30799908

fbshipit-source-id: 403b817a7c72cc747e88ef6eb98f9313a3e3ca1a