StringValue type
Summary:
`StringValue` is a wrapper to `Value` which asserts that contained value is `StarlarkStr`.

(Similar to `FrozenStringValue` which is logically a wrapper for `FrozenValue` which asserts that contained value is `StarlarkStr`).

Used in the following diff D30884359.

In theory `StringValue` can be used for faster operations on strings (e. g. obtaining a string from `StringValue` is cheaper than obtaining a string from `Value`), but there's no obvious applications for that at the moment.

Reviewed By: ndmitchell

Differential Revision: D30884361

fbshipit-source-id: 21b632f2a6a4362eb05dda35b0e2a12162397faf