Inline funtions like is_list
Summary:
Inline functions like:

```
def is_list(x):
  return type(x) == type([])
```

Other function can be inlined too, but more work required to do that (for example, we know that `is_list` function is infallible, but to inline fallible functions, stack trace need to be preserved).

Note starlark-java (in Buck v1) has this optimization.

I also have a stack for Buck v1 to inline any short functions, D28611893, but it was not committed.

Reviewed By: ndmitchell

Differential Revision: D30888372

fbshipit-source-id: fe3d4e0361efe2100542fa99a2dcd337e7a16b41