Change Display for Value strings to be "quoted"
Summary:
Before Display on a string gave us the string without quotes, now it includes them. This is important if we want repr to be equivalent to Display, which seems like a very good default. Therefore, add quotes to Display.

Importantly, any code using to_string on a string, which expected the raw string, now gets quotes. In most cases we shouldn't be calling to_string as its more expensive than unpack_str (by a lot), but might be some cases.

Reviewed By: stepancheg

Differential Revision: D31032072

fbshipit-source-id: b05a1ea66694e4ab746ef1da47e2a8a1d63d685f