buck2: add support for emitting heap utilization in a flame graph
Summary:
Like it says in the title. The underlying goal is to let us use tools like
flamegraph.pl when looking at an analysis profile. This uses the same data as
the current summary, just presented differently.

Note that this doesn't reuse a comparable implementation we have in CPU
profiling. The reason is they have different goals (this one doesn't need to be
efficient to GC, which is a pretty big difference, but also they ingest different
input).

It's not wired up to Buck yet (or exposed), that's coming next!

Reviewed By: ndmitchell

Differential Revision: D31022233

fbshipit-source-id: 15d36f85d1f55c346f133a8e24ff5179a9d5b364