Use FrozenRef<FrozenModuleRef> instead of FrozenModuleValue
Summary:
Before this diff:

```
FrozenModuleValue(FrozenValue)
```

After previous diff that `FrozenValue` was always `FrozenModuleRef` and not a `BlackHole`. So there's no longer need to store it as frozen value and do downcast.

So just remove `FrozenModuleValue` and replace this type was `FrozenRef<FrozenModuleRef>`.

Frozen function call no longer needs to dynamically downcast `FrozenValue` to `FrozenModuleRef`.

Reviewed By: ndmitchell

Differential Revision: D31003787

fbshipit-source-id: c451068fbf416dcd460f0df5f265eea9f7b982b4