Do not take ExprCompilerValue by value in as_compiled
Summary:
Change:

```
fn as_compiled(self) -> ExprCompiledValue { ... }
```

to

```
fn as_compiled(&self) -> ExprCompiledValue { ... }
```

We need to preserve original `ExprCompiledValue` after compilation to do the compilation again on freeze.

Unfortunately, this diff results in many string duplications at compile time. For example, in "get local" expression we need string to use it in error message. Now we clone the string.

This can be addressed for example by
* placing all strings into frozen heap (less memory safe)
* placing string in `Arc` (almost as expensive as string copy)
* creating some string table per module and store indices to strings (too complicated)

It string copy is the issue, perhaps the best thing to do is to stop allocating strings at parse time: we already have spans for the strings, so we can store strings as spans through the pipeline.

Reviewed By: ndmitchell

Differential Revision: D31034896

fbshipit-source-id: 8ea4776f3f8493cc6bd72ef27be8e3f564491dde