Ensure all tests derive Display for anything StarlarkValue
Summary: For the docs, the choices are to use display_more (simpler, but maybe involves something the user isn't familiar with). Or to write it out manually. Given the manual variant significantly overshadows what I want the docs to say, I've gone with derive_more, apart from the initial example on the front page, where clarity matters more.

Reviewed By: bobyangyf

Differential Revision: D31303178

fbshipit-source-id: a9a6bc62423ae9e708ca1f144c0afc1611448f58