Default collect_repr to Display, not Debug
Summary:
Generally repr should be carefully chosen, since it shows up everywhere, and Display is a better default than Debug.

This is a breaking change for anything that had different Display and Debug, but didn't implement collect_repr. The main one this impacts is things like StarlarkAny containing a String (hence the test changes).

Reviewed By: bobyangyf

Differential Revision: D31304119

fbshipit-source-id: 3d92726bda6b564cfd03b7f6155b397f7ee5194b