Use #[thread_local]
Summary:
`#[thread_local]` is not supported everywhere (some INeverHeardOfBSD probably) but we are not using these platforms.

AFAIU it is supported on Linux, macOS and Windows.

This diff replaces `thread_local!` macro with `#[thread_local]` attribute.

Alternatively, we can keep existing implementation and switch to `#[thread_local]` only when `cfg(target_thread_local)`.

[Rust issue about #[thread_local] attribute](https://github.com/rust-lang/rust/issues/29594).

Reviewed By: ndmitchell

Differential Revision: D31415168

fbshipit-source-id: 923b27084ac941bb9a326f83c43a9701e850fe65