Fix memory corruption in alloca
Summary:
Pointers are corrupted after `alloca_more`.

We never encountered it because we allocate 1Mb initially, so practically we never grow the alloca buffer.

Currently max recursion is 40. If we have 1000 parameters per function in call stack of 40, then we only need 320K of memory to store all these parameters.

This fix adds more overhead to `alloca` (one more register to save).

Alternatively to this diff we can just rewrite alloca with non-growable buffer.

Reviewed By: ndmitchell, krallin

Differential Revision: D31467038

fbshipit-source-id: 163deba088fb097105fc691a01df2f8528e8c809