Bytecode interpreter
Summary:
This diff replaces closure interpreter with bytecode interpreter.

# Bytecode

Instruction is opcode + arguments. Each instruction for given opcode has fixed size.

Instruction pointer is a memory pointer to instructions buffer.

The interpreter loop looks like this:

```
fn run_block(mut ip: *const u8) -> Control {
  loop {
    let opcode = *ip;
    ip = match opcode {
      Add => {
        pop2();
        add();
        push();
        ip + sizeof::<InstrAdd>()
      }
      ...
      Return => return Return,
      Continue => return Continue,
    }
  }
}
```

Note there's no range check here, bytecode block must end with either `Return` or `Continue` instruction.

`run_block` function is fully inlined down to virtual functions, so the whole interpreter is one big function in the machine code.

# Stack vs register

This diff implement stack interpreter.

Unlike Starlark in Buck v1 which implements register interpreter.

**The advantage of register interpreter** is that it allows unifying local variables and temporaries, i. e. all instructions accept arguments by index.

**The advantage of stack interpreter** is that when function is called, arguments can be passed just by taking a pointer to the stack. This was my main reason for choosing the stack interpreter. This can be also done with register interpreter, but it is trickier.

# Verbose code

The instruction implementation is quite verbose. With this implementation each instruction is an implementation of a trait:

```
trait BcInstr {
  type Pop; // values to pop off the stack
  type Push; // values to push to the stack
  type Arg; // fixed instruction argument

  fn invoke(.., arg: &Self::Arg, pop: Self::NPop) -> Push;
}
```

so most instructions are safe on both sides (serialization and deserialization).

Unfortunately it leads to quite verbose code due to limitation of rust type system, e. g. it is not possible to parameterize type with closure (at least I have not found a good way to do it).

On the other hand, this diff adds 5KLOC, while `eval.fragment` is 4KLOC (most of it should be removed), so it is not a big regression in code size.

# Unsafety

The interpreter does unsafe code, mostly to
* get rid of range checks
* use a pointers directly instead of base+offset indirection

Many functions marked safe are actually unsafe (but they are not visible outside of `bc` module).

# Optimizations included here

This diff intentionally doesn't do any optimizations not specific to bytecode.

Bytecode specific optimizations included are:

## ConstN, LoadLocalN instructions

`ConstN` instruction puts multiple values to the stack.

`LoadLocalN` instruction reads multiple local variables in single instruction.

## if and

```
if a and b { ... }
```

with closure interpreter is compiled as:

```
fn c() { a() and b() }
if c() { ... }
```

With bytecode interpreter it is compiled as:

```
if a() {
  if b() {
    ...
  }
}
```

However optimization for `if a or b { ... }` is not included here because it is harder to implement.

# Future optimizations

The most important missing feature here is instruction profiler.

We need to collect statistics of instruction count/time by opcode. Buck v1 Starlark does that.
* Most frequent instructions could to be split into longer instructions
* Infrequent instructions might be need to be merged into more generic instructions

For example, this diff has `Const2`, `Const3`, `Const4` instructions, but the number 4 was chosen arbitrarily. Maybe we don't need `Const4`, or maybe we need `Const5` and `Const6`.

# Threaded interpreter?

Some people suggest doing threaded/CPS style interpreter.

My understanding is threaded interpreter works well only for simple instructions like add two integers. If there are calls inside an instruction, generated code starts [saving register](https://rust.godbolt.org/z/1Gss7PWEf), which is not free. My experiments with doing threaded interpreter made from this diff led to slowdown.

# How much faster it can be?

Currently `run_block` function, which is the bytecode interpreter loop, takes 13% of total CPU time (excluding nested calls). Which means that if the interpreter core (closure, bytecode, machine code or whatever) could be infinitely fast, overall speedup would be 13%.

This diff saves 4% of total execution time. So it means, the interpreter core (part which evaluates instructions) is now `4/13 ~= 30%` faster.

Reviewed By: ndmitchell

Differential Revision: D31340035

fbshipit-source-id: 4749b7e71052ef780620d5ebb13318fcdde9749f