Fix repr of StarlarkStrNRepr
Summary:
New struct hierarchy:

```
struct StarlarkStrN<N> {
  len: usize,
  content: [u8; N],
}

struct StarlarkStr(StarlarStrN<0>);

struct StarlarkStrNRepr<N>(AValueRepr<StarlarkStrN<N>>);
```

This fixes the issue when offset of string content is not equal to the size of `StarlarkStr`. Which is the case for example with this string representation:

```
struct StarlarkStr {
  hash: u64,
  len: u32,
  content: [u8; 0],
}
```

Here offset of `content` is 12, but size of `StarlarkStr` is 16.

Reviewed By: ndmitchell

Differential Revision: D31593184

fbshipit-source-id: b7950748d8632c9c6e6b7ce598fe75373b863f78