Test hashes are consistent between Value and StringValue
Summary:
However, we should not use `Hash` trait of `StringValue` to produce a dict or struct key, because it misses the chance of reusing `StarlarkStr` cached hash.

(That is not a regression, this is how `alloc_str_hashed` worked before this stack: it didn't cache hash either.)

I'll address that later.

Reviewed By: ndmitchell

Differential Revision: D31616672

fbshipit-source-id: a4eff514caa3417c1efd82507d46def37e7f460f