#[trace(unsafe_ignore)]
Summary:
`#[trace(unsafe_ignore)]` annotation is used with `#[derive(Trace)]` and can be applied to a struct field or enum variant.

When applied to a field or variant, `starlark_derive` skips annotated field/enum variant in generated `Trace` implementation.

`#[trace(unsafe_ignore)]` saves some lines of code, but also it makes code safer, because when adding fields, with this annotation new field need to be explicitly annotated, and with manual implementation it is easy to forget to trace new field.

Reviewed By: ndmitchell

Differential Revision: D31643108

fbshipit-source-id: 2484c1c3b58bcd11c60c8227fe0a5db0b5d289f3