read_line abstraction
Summary:
Thin wrapper for rustyline.

We use `rustyline` in two places: in REPL and in `breakpoint` function. Use the common code to call `rustyline` to make sure that both of them work correctly.

Alternatively we can kick `rustyline` out of `starlark` library, and always rely on users providing `BreakpointConsole` callback.

Reviewed By: ndmitchell

Differential Revision: D31676197

fbshipit-source-id: b042060f32189ac758c0f7663f9e0740ba66e18e