Add OwnedFrozenValue::default
Summary:
Useful to have, with the same default as for Value/FrozenValue.

Only choice was whether to create a single Arc and increment it, or create a fresh one each time. I went for a fresh one, since I suspect the memory usage is pretty low, but could revisit later.

Reviewed By: stepancheg

Differential Revision: D31769608

fbshipit-source-id: a9863b092c39b185da0eb344f8ae32ac772b17b5