Add spans in derive impl
Summary:
Before:

```
error[E0759]: `eval` has lifetime `'v` but it needs to satisfy a `'static` lifetime requirement
  --> starlark-rust/starlark/src/assert/assert.rs:98:1
   |
98 | #[starlark_module]
   | ^^^^^^^^^^^^^^^^^^
   | |
   | this data with lifetime `'v`...
   | ...is captured and required to live as long as `'static` here
   |
   = note: this error originates in the attribute macro `starlark_module` (in Nightly builds, run with -Z macro-backtrace for more info)
```

Now:

```
error[E0759]: `eval` has lifetime `'v` but it needs to satisfy a `'static` lifetime requirement
   --> starlark-rust/starlark/src/assert/assert.rs:101:11
    |
101 |       fn eq(a: Value, b: Value) -> NoneType {
    |          -  ^^^^^^^^^^^^^^^^^^ ...is captured and required to live as long as `'static` here
    |  ________|
    | |
102 | |         assert_equals(a, b)
103 | |     }
    | |_____- this data with lifetime `'v`...
```

Not always precise (and sometimes it is not possible to be precise), but at least it is clear now which function triggers error.

This diff does not affect errors in function bodies, which were highlighted correctly before and highlighted correctly after this diff.

Reviewed By: ndmitchell

Differential Revision: D31866390

fbshipit-source-id: 04a714600b3eae8f213d12e03501ed90851c808e