Print out errors less
Summary:
Before, if Starlark produced an error (e.g. fail) then error would appear twice. It appeared once when we displayed the diagnostic, and once when we returned the underlying error as the `source` from `anyhow::Error` which was printed with `{:#}`.

I'm having difficulty threading the needle of when to use source. It seems like the answer might only be when we wrap the error, but the wrapping doesn't print out the underlying piece? Either way, this should reduce the repeated errors problem.

Reviewed By: stepancheg

Differential Revision: D31971408

fbshipit-source-id: 6a5e5dd4aa814902392bce04c6e0035be6e2b833