List's `iter` can be `ExactSizeIterator` (#41)
Summary:
I'm toying around with making PyO3 bindings for this library and I found this helpful going from a Starlark `List` -> `PyList`.

It seems like a reasonable change since lists always know their length, and due to the use of `content.iter` there's no additional implementation burden. `ExactSizeIterator` is a strictly more confined type than `Iterator` so this is fully backwards compatible. I'm not sure if this can be applied to generators, but I should think not as an error could occur during iteration? Presumably this same change could be made to tuples and other types quite easily.

I'll get to those eventually in my gacially slow just-for-cruriosity's-sake pace, but wanted to open a PR now, as I'm not sure how often I'll tinker on this project.

Cheers

Pull Request resolved: https://github.com/facebookexperimental/starlark-rust/pull/41

Reviewed By: stepancheg

Differential Revision: D32049025

Pulled By: ndmitchell

fbshipit-source-id: 7485ce96cbbb3bbc5b4b68580859ecb167436fd0