Split Globals
Summary:
Extract `Methods` from `Globals`, `MethodsBuilder` from `GlobalsBuilder`.

* `Globals` is globals passed when evaluating a module
* `Methods` is what returned from `StarlarkValue::get_methods`

`Globals` and `Methods` are similar but not the same: `Globals` members must not accept `this` while `Methods` members must accept `this`. If we split `Globals` we should be able to verify that at compile time (not in this diff).

Also, IMHO this makes code a little more readable: looking at function annotated with `#[starlark_module]` it is immediately clear, if it is globals or methods.

The downside is copy-paste.

Reviewed By: ndmitchell

Differential Revision: D32129431

fbshipit-source-id: fad820abdf279e5dbd5ba4efc76e51880abbbf9b