spec_safe attr
Summary:
`#[starlark(spec_safe)]` function attribute.

When a function is `spec_safe` the function is safe to be executed speculatively, which means it should:
* not panic
* not run for too long and don't consume too much memory (otherwise speculative execution may have negative effect on compilation time on invocation of functions which are never executed)
* have no side effects

Spec-safe functions are allowed to return mutable non-frozen values, but the optimizer won't always be able to use results of such function invocations.

Following diff D32186393 optimizes evaluation using this attribute.

Reviewed By: ndmitchell

Differential Revision: D32186394

fbshipit-source-id: 50f6c44e90869b0d52026f239fa4c5b49f9eba15