Add a StarlarkValue for str
Summary: Often things want to reuse the "string version" of something in StarlarkValue. Previously that was done with custom methods (e.g. json_string), but it's annoying to reuse like that. Instead, introduce StarlarkValue str, which StarlarkValue StarlarkStr delgates to in most cases. No functional change, but now you can do collect_repr on a str, for example.

Reviewed By: stepancheg

Differential Revision: D32238138

fbshipit-source-id: 9f86241c576e8970e3291659644c0a5ecde9dd3f