Remove newlines in docstring summaries
Summary:
There are some places that we want to use these summaries, where the line breaks can cause issues. We always want summaries to be a single line, so just remove their newlines.

(Example place this breaks: Trying to print the summary in a markdown table if it has newlines)

Reviewed By: ndmitchell, stepancheg

Differential Revision: D32109406

fbshipit-source-id: df531a49552c3ebff816f0cb1531511af4a70222