Fix iter_params off-by-some error
Summary: Some built in functions, like `min()`, take "*args" as the "first" param, and have a second, named parameter ("key" in min's case). This causes off-by-one errors, because of an implicit assumption that *args/**kwargs are the last ones in the list. Just change our `names` idx based on whether we've see args/kwargs.

Reviewed By: ndmitchell, stepancheg

Differential Revision: D32302388

fbshipit-source-id: 63adc54649cf7242354f2b0ae16748759c8d2631