Make sure to document args for positional-only functions
Summary: Before, because we don't create signatures in order to call positional-argument-only functions, we weren't getting any parameter details for functions like `min()`, or `range()` that have optional/positional only arguments. Add in a function to get a signature for those, too, and move the documentation generation logic in render() to its own function, as it was getting a bit big.

Reviewed By: bobyangyf

Differential Revision: D32364984

fbshipit-source-id: 93bae477d263463ff355b7e91bd55cfd5e2d0c3e