Expose documentation for each member in globals
Summary: Generally we want to use `documentation()` to get docs in a format that looks like documentation for a specific object type. However, other times we might want to load a number of symbols into a `GlobalsBuilder` and grab the documentation for each of those individually. Expose a method that maps each symbol to its `DocItem` object.

Reviewed By: stepancheg

Differential Revision: D32438987

fbshipit-source-id: a2693e5b8165f6f73e8aec693ccfddfa1dcf561c