BcInstrSlowArg
Summary:
in the end of bytecode we store span for each instruction.

Span is needed only for errors, so storing it in the end of bytecode (instead of next to instruction as for regular instruction arguments) makes accessing it very slow, but normal execution of instruction a little faster.

This diff generalizes this span into "slow instruction argument": anything which instruction need infrequently.

With this diff, `BcInstrArg` only contains instruction span.

Following diff D32874230 adds `Vec<Span>` to `BcInstrArg` for instructions which need multiple spans (like `LoadLocal3`).

Reviewed By: ndmitchell

Differential Revision: D32874152

fbshipit-source-id: bf0a0734e951fbbda811d3031cbe3b7c8050d8b5