Make Span type package private
Summary:
`Span` is hard to use correctly: it is not tied to `CodeMap`.

Another reason to get rid of `Span` from public API is to be able to implement function inlining: inlining may be good for performance, but inlined functions may belong to a different file, so have different `CodeMap`. When `Span` is no longer public, such changes are easier to make.

Reviewed By: ndmitchell

Differential Revision: D32875024

fbshipit-source-id: e6bc1f942efec1a2dc439cfbab1cc3e7307a1144