Change how bytecode frame is allocated
Summary:
Before this diff frame data (locals + stack) was allocated in alloca, but frame header (local count + stack size + pointer to data) was stored in `Evaluator`. When entering new frame, previous frame header was preserved in native frame.

Now `BcFrame` is DST (frame data is followed by `BcFrame` struct). Full `BcFrame` with data is allocated in alloca. And `Evaluator` now contains a pointer to `BcFrame` (called `BcFramePtr` now).

This change is prerequisite to implementation of proper GC: we need to store the stack of frames to be able to trace locals and stack values. Before this diff it was somewhat not trivial.

Now it is possible to naturally store a pointer to the parent frame when allocating a new frame. So all frames would be a linked list.

Reviewed By: ndmitchell

Differential Revision: D32519506

fbshipit-source-id: 79bde4ef997a033e119c40b803a75c5075e13c95