Faster symbol map lookup on negative
Summary:
Consider this code:

```
foo(**kwargs): ...

d = {"a": 1}
foo(**d)
```

Here we lookup the string "a" in parameter names map of `foo`.

When performing lookup, we construct `&str` from `StringValue`. Which is dereference of string to fetch length.

But we don't need to dereference when hashmap lookup is definitely negative like in the example above.

So postponing this dereference speeds up negative look up a little.

Reviewed By: ndmitchell

Differential Revision: D32938022

fbshipit-source-id: 52c7fb03f81c23cde9c88c588318f4a211d9a43d