Rename the custom_linter feature to gazebo_linter
Summary: When it was named custom_linter, we hadn't yet named our linter, so it makes sense for them to be consistent. I was tempted to call it gazebo_lint as a feature, but you can only have features with the same names as dependencies in nightly, so gazebo_linter will have to do (and also makes it more grep'able).

Reviewed By: krallin

Differential Revision: D34374861

fbshipit-source-id: 8a623600546fc81b8dfb936cc84502b983266e27