starlark-rust: add some tests for Freeze derive
Summary:
Like it says in the title. I added a bit of non-trivial functionality in the
previous diffs so at this point I'd like to make sure it has some test
coverage.

Note that this has to live in the Starlark crate because the code generated by
Starlark derive depends on it, and because creating a Freezer in tests is a lot
easier if you have access to the `pub(crate)` APIs.

Reviewed By: stepancheg

Differential Revision: D34577562

fbshipit-source-id: 6003c8b83a08cc7bd4aa74cf0d4fc8976a140054