starlark-rust: derive: allow setting bounds on the Freeze implementation
Summary:
Like it says in the title. This makes the validator introduced in the diff more
usable, since it lets you place bounds on the types (like `V: ValueLike`) that
your validator is going to look at.

Note that I've also introduced a `'freeze` lifetime on the implementation that
users can leverage in their bounds (this is similar to e.g. what `async_trait`
does with `'async_trait`).

Reviewed By: stepancheg

Differential Revision: D34577561

fbshipit-source-id: 0843169a31ec26a3dff7e1027235d14651c0394d