downcast unchecked in comprehension instrictions
Summary:
For compressions are implemented by placing a result collection to the stack and appending an element to it in the end of each iteration, in instructions `ComprListAppend` and `ComprDictInsert`.

This diff patches these two instructions to assume the stack always contains the value of corresponding type instead of doing safe cast: if we generate bytecode correctly, this is always the case.

Reviewed By: ndmitchell

Differential Revision: D33743809

fbshipit-source-id: b6fb9722668a5d063d259f51d320892682183d51